[AOS] feat: 트림 선택 페이지 API 연동 및 기존 버그 수정 (#273)
* [AOS] feat: api 통신 부 코드 수정 #262

* [AOS] refactor: 불 필요한 코드 제거 #262

* [AOS] fix: 가격 타입 변경 #262

* [AOS] fix: 아이템 클릭시 blinking issue 해결 #262

* [AOS] feat: api 데이터 연동 #262

* [AOS] refactor: 데이터 바인딩 코드 개선 #262

* [AOS] feat: 트림 상세 설명 api 연동 #262

+++ Android/app/src/main/java/org/softeer_2nd/caArt/view/bindingAdapter/TextViewBindingAdapter.kt
@@ -26,7 +26,7 @@ fun TextView.setTextSizeWithFloat(size: Float) {
 }
 
 @BindingAdapter("formattedPrice")
-fun TextView.setFormattedPrice(price: Int) {
+fun TextView.setFormattedPrice(price: Long) {
     val formatter = NumberFormat.getNumberInstance(Locale.KOREA)
     val formattedPrice = formatter.format(price)
     text = context.getString(R.string.formatted_price, formattedPrice)
